Commiting the first creation of TerrainGeneration.ca
Terrain generation contains the functionality to build the map using Perlin Noise functionality. Using a tile system it spawns grass prefabs as well as the Tower prefab before generating the paths. It keeps track of important variables such as the starting points on each path to be used in enemy spawning and created a defender area for defenders to be places later. It bakes a navmesh surface onto the path to be implemented in the enemy movement functionailty